% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kotlarski.R
\name{cf2dens.inner}
\alias{cf2dens.inner}
\title{cf2dens.inner}
\usage{
cf2dens.inner(t, cf, x)
}
\arguments{
\item{t}{particular value to evaluate characteristic function at}

\item{cf}{a characteristic function (should be evaluated at t)}

\item{x}{a particular value for the pdf}
}
\value{
scalar imaginary number
}
\description{
a helper function for computing the inverse Fourier transform
 of a characteristic function.
}
