% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kotlarski.R
\name{cf2dens}
\alias{cf2dens}
\title{cf2dens}
\usage{
cf2dens(cf, tgrid, xgrid)
}
\arguments{
\item{cf}{a characteristic function (will be evaluated at all values of
tgrid)}

\item{tgrid}{vector of all possible values to evaluate characteristic
function}

\item{xgrid}{vector of values to recover pdf of corresponding characteristic
function cf}
}
\value{
vector of values of pdf for all values of xgrid
}
\description{
compute inverse Fourier transform of a characteristic function
 (i.e., turn a characteristic function back into a pdf)
}
