% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kotlarski.R
\name{cfE}
\alias{cfE}
\title{cfE}
\usage{
cfE(s, lam = 1)
}
\arguments{
\item{s}{value to evaluate the characteristic function at (can be vector)}

\item{lam}{rate parameter}
}
\value{
vector of the values of the characteristic function for each
 value of s
}
\description{
characteristic function of exponential(\lambda)
}
