% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kotlarski.R
\name{cfN}
\alias{cfN}
\title{cfN}
\usage{
cfN(s, mu = 0, sig2 = 1)
}
\arguments{
\item{s}{value to evaluate the characteristic function at (can be vector)}

\item{mu}{mean}

\item{sig2}{variance}
}
\value{
vector of the values of the characteristic function for each
 value of s
}
\description{
characteristic function of N(\mu,\sigma^2)
}
