% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kotlarski.R
\name{cfX}
\alias{cfX}
\title{cfX}
\usage{
cfX(tvec)
}
\arguments{
\item{tvec}{vector of values to evaluate the characteristic function at}
}
\value{
vector of the values of the characteristic function of X for
 each value of tvec
}
\description{
Estimate the characteristic function of some random variable X
}
