% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kotlarski.R
\name{cfX2}
\alias{cfX2}
\title{cfX2}
\usage{
cfX2(t1, t2)
}
\arguments{
\item{t1}{evaluate characteristic function at this value}

\item{t2}{evaluate characteristic function at this value}
}
\value{
scalar value of characteristic function at t1 and t2
}
\description{
characteristic function with two variables
}
