% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kotlarski.R
\name{cfXX}
\alias{cfXX}
\title{cfXX}
\usage{
cfXX(t)
}
\arguments{
\item{t}{value to evaluate the characteristic function at (can be vector)}
}
\value{
vector of the values of the characterstic function of X
 for each value of t
}
\description{
distribution of X when e_1 is known to be ~N(0,1)
}
