% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kotlarski.R
\name{d1cfX2}
\alias{d1cfX2}
\title{d1cfX2}
\usage{
d1cfX2(t1, t2)
}
\arguments{
\item{t1}{evaluate derivative at this value}

\item{t2}{evaluate derivative at this value}
}
\value{
scalar value of derivative of characteristic function at t1 and t2
}
\description{
derivative of characteristic function with two variables
 with respect to first variable
}
