% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kotlarski.R
\name{kot.inner}
\alias{kot.inner}
\title{kot.inner}
\usage{
kot.inner(s)
}
\arguments{
\item{s}{a particular value to evaluate the charactersitic function at}
}
\value{
scalar value of characteristic function
}
\description{
kotlarski theorem result for single value of s;
 that is, estimate the characteristic function of the true X
 with two measures X_1 = X + e_1, X_2 = X + e_2.  This function is
 called by a wrapper that obtains the kotlarski result for a
 vector of values of s
}
