% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kotlarski.R
\name{kotlarski}
\alias{kotlarski}
\title{kotlarski}
\usage{
kotlarski(tvec)
}
\arguments{
\item{tvec}{a vector of values to estimate the characteristic function at}
}
\value{
vector of the values of the characteristic function of X
 for each value of tvec
}
\description{
Estimate the characteristic function of the true X
 with two measures X_1 = X + e_1, X_2 = X + e_2
}
